Added python unit testing

+++ pysiglib/pysiglib.py
@@ -1,9 +1,10 @@
 import numpy as np
 import ctypes
-from ctypes import c_double, c_int, c_bool, POINTER
+from ctypes import c_double, c_int64, c_bool, POINTER
 import os
 import sys
 import time
+import dataHandling
 
 dir_ = os.path.dirname(sys.modules['pysiglib'].__file__)
 print(dir_)
@@ -26,14 +27,14 @@ def getPathElement(numpyArray, lengthIndex, dimIndex):
     dataDimension = numpyArray.shape[1]
     dataPtr = numpyArray.ctypes.data_as(POINTER(c_double))
 
-    cpsig.getPathElement.argtypes = (POINTER(c_double), c_int, c_int, c_int, c_int)
+    cpsig.getPathElement.argtypes = (POINTER(c_double), c_int64, c_int64, c_int64, c_int64)
     cpsig.getPathElement.restype = c_double
 
     return cpsig.getPathElement(dataPtr, dataLength, dataDimension, lengthIndex, dimIndex)
 
 def polyLength(degree, dimension):
-    cpsig.polyLength.argtypes = (c_int, c_int)
-    cpsig.polyLength.restype = c_int
+    cpsig.polyLength.argtypes = (c_int64, c_int64)
+    cpsig.polyLength.restype = c_int64
     return cpsig.polyLength(degree, dimension)
 
 def signature_(path, degree, timeAug = False, leadLag = False, horner = True):
@@ -42,17 +43,17 @@ def signature_(path, degree, timeAug = False, leadLag = False, horner = True):
     out = np.empty(shape=polyLength(dimension, degree), dtype=np.float64)
 
     if np.issubdtype(path.dtype, np.integer):
-        dataPtr = path.ctypes.data_as(POINTER(c_int))
-        outPtr = out.ctypes.data_as(POINTER(c_int))
+        dataPtr = path.ctypes.data_as(POINTER(c_int64))
+        outPtr = out.ctypes.data_as(POINTER(c_int64))
 
-        cpsig.signatureInt.argtypes = (POINTER(c_int), POINTER(c_int), c_int, c_int, c_int, c_bool, c_bool, c_bool)
+        cpsig.signatureInt.argtypes = (POINTER(c_int64), POINTER(c_int64), c_int64, c_int64, c_int64, c_bool, c_bool, c_bool)
         cpsig.signatureInt(dataPtr, outPtr, dimension, length, degree, timeAug, leadLag, horner)
 
     elif np.issubdtype(path.dtype, np.floating):
         dataPtr = path.ctypes.data_as(POINTER(c_double))
         outPtr = out.ctypes.data_as(POINTER(c_double))
 
-        cpsig.signature.argtypes = (POINTER(c_double), POINTER(c_double), c_int, c_int, c_int, c_bool, c_bool, c_bool)
+        cpsig.signature.argtypes = (POINTER(c_double), POINTER(c_double), c_int64, c_int64, c_int64, c_bool, c_bool, c_bool)
         cpsig.signature(dataPtr, outPtr, dimension, length, degree, timeAug, leadLag, horner)
 
     else:
@@ -67,17 +68,17 @@ def batchSignature_(path, degree, timeAug = False, leadLag = False, horner = Tru
     out = np.empty(shape=(batchSize, polyLength(dimension, degree)), dtype=np.float64)
 
     if np.issubdtype(path.dtype, np.integer):
-        dataPtr = path.ctypes.data_as(POINTER(c_int))
-        outPtr = out.ctypes.data_as(POINTER(c_int))
+        dataPtr = path.ctypes.data_as(POINTER(c_int64))
+        outPtr = out.ctypes.data_as(POINTER(c_int64))
 
-        cpsig.batchSignatureInt.argtypes = (POINTER(c_int), POINTER(c_int), c_int, c_int, c_int, c_int, c_bool, c_bool, c_bool, c_bool)
+        cpsig.batchSignatureInt.argtypes = (POINTER(c_int64), POINTER(c_int64), c_int64, c_int64, c_int64, c_int64, c_bool, c_bool, c_bool, c_bool)
         cpsig.batchSignatureInt(dataPtr, outPtr, batchSize, dimension, length, degree, timeAug, leadLag, horner, parallel)
 
     elif np.issubdtype(path.dtype, np.floating):
         dataPtr = path.ctypes.data_as(POINTER(c_double))
         outPtr = out.ctypes.data_as(POINTER(c_double))
 
-        cpsig.batchSignature.argtypes = (POINTER(c_double), POINTER(c_double), c_int, c_int, c_int, c_int, c_bool, c_bool, c_bool, c_bool)
+        cpsig.batchSignature.argtypes = (POINTER(c_double), POINTER(c_double), c_int64, c_int64, c_int64, c_int64, c_bool, c_bool, c_bool, c_bool)
         cpsig.batchSignature(dataPtr, outPtr, batchSize, dimension, length, degree, timeAug, leadLag, horner, parallel)
 
     else:
@@ -95,3 +96,5 @@ def signature(path, degree, timeAug = False, leadLag = False, horner = True, par
 
 
 #https://stackoverflow.com/questions/64478880/how-to-pass-this-numpy-array-to-c-with-ctypes
+
+#gpu ptr might have to be c_void_p
